# docs/swagger.yaml
basePath: /
definitions:
...
  history-api_internal_dtos_request.CreateUserVerificationDto:
    properties:
      content:
        minLength: 10
        type: string
      media_ids:
        items:
          type: string
        type: array
      verify_type:
        enum:
        - ID_CARD
        - EDUCATION
        - EXPERT
        - OTHER
        type: string
    required:
    - content
    - verify_type
    type: object
...
  history-api_internal_dtos_request.UpdateVerificationStatusDto:
    properties:
      review_note:
        maxLength: 3000
        minLength: 5
        type: string
      status:
        enum:
        - PENDING
        - APPROVED
        - REJECTED
        type: string
    required:
    - review_note
    - status
    type: object
...
  history-api_internal_dtos_response.CommonResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for History API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: History API
  version: "1.0"
paths:
...
  /historian/application:
    get:
      consumes:
      - application/json
      description: Get list of historian applications with filters
      parameters:
      - in: query
        name: created_from
        type: string
      - in: query
        name: created_to
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        name: reviewed_by
        type: string
      - in: query
        maxLength: 200
        minLength: 2
        name: search
        type: string
      - enum:
        - id
        - created_at
        - reviewed_at
        - status
        in: query
        name: sort
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: statuses
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: user_ids
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: verify_types
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Search historian applications
      tags:
      - Historian Application
    post:
      consumes:
      - application/json
      description: Submit application to become historian
      parameters:
      - description: Application data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.CreateUserVerificationDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Create historian application
      tags:
      - Historian Application
  /historian/application/{id}:
    delete:
      description: Delete historian application
      parameters:
      - description: Verification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Delete application
      tags:
      - Historian Application
    get:
      description: Get historian application detail
      parameters:
      - description: Verification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get application by ID
      tags:
      - Historian Application
  /historian/application/{id}/status:
    put:
      consumes:
      - application/json
      description: Approve or reject historian application
      parameters:
      - description: Verification ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.UpdateVerificationStatusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Update application status
      tags:
      - Historian Application
...
  /wikis:
    get:
      consumes:
      - application/json
      description: Search wikis with cursor pagination
      parameters:
      - in: query
        name: cursor
        type: string
      - in: query
        name: entity_id
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maxLength: 1000
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      summary: Search wikis
      tags:
      - Wikis
  /wikis/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific wiki
      parameters:
      - description: Wiki ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      summary: Get wiki by ID
      tags:
      - Wikis
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"